Removed _copy_over_loops flag and moved the base extractor and writer to the common namespace
