+++ app/services/BaseExtractorService.py
