+++ app/services/marker/serato/v2/MarkerExtractorService.py
@@ -12,7 +12,7 @@ from app.serializers.serato.v2.BpmLockSerializer import BpmLockSerializer
 from app.serializers.serato.v2.ColorSerializer import ColorSerializer
 from app.serializers.serato.v2.CueSerializer import CueSerializer
 from app.serializers.serato.v2.LoopSerializer import LoopSerializer
-from app.services.marker.BaseExtractorService import BaseExtractorService
+from app.services.BaseExtractorService import BaseExtractorService
 from app.utils.serato import read_bytes
 from app.utils.serato.type_detector import detect_type
